test: split 3 remaining oversized files under 500 lines
Split d1-queries.search.test.ts, mcp-tools-execution.test.ts, and
pipeline-lock-maintenance.test.ts which were still over 500 lines.
All new split files pass tests (78 tests total).
